requirements.txt

fastapi
uvicorn
python-dotenv
requests
msal
transformers
botbuilder-core
aiohttp
